requirements.txt
fastapi
uvicorn
python-multipart
pydantic
httpx
slowapi
python-dotenv
